requirements.txt
# Gold Intelligence Platform - Requirements
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Data
pandas==2.2.3
polars==1.8.0
numpy==1.26.4
pyarrow==17.0.0

# ML / Modeling
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.1
catboost==1.2.7
hmmlearn==0.3.2
shap==0.46.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.0

# Data Sources
requests==2.32.3
httpx==0.27.2
fredapi==0.5.2
newsapi-python==0.2.7

# Scheduling / Tasks
celery==5.4.0
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==9.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0

# Visualization / Reports
plotly==5.24.1
matplotlib==3.9.2
jinja2==3.1.4
